Specify that each slot is an OWL annotation property and assertion
This is a temporary solution as it may fail if a same field is used by both the metadata object and data instances. For instance, we didn't state that 'title' (in metadata-base.yaml) is an OWL annotation property and assertion because it's used by the metadata object (see normalized.yaml). If a data instance wants to use the 'title' property then we need to add the OWL annotation property and assertion definitions in its schema file. You can see this workaround in the collection-metadata.yaml.